Fix the bug when game_speed is _too_ fast (500K faster than real time) it was causing the produced events from schedule shift into the out of order timestamps.
I feel that depending on the fraction might break somewhere else in future;
can we just rely on the ever incremental int to do the same?
Right now, the since Tick generation relies on the real timestamp diff
to produce the number of ticks we want at any given point in time.
This kind of blocks changing that easily.